Fix type.
CVS patchset: 4347
CVS date: 2004/11/17 18:59:43

+++ lib/clplumbing/cl_msg_types.c
@@ -287,7 +287,7 @@ binary_dup(const void* value, size_t len){
 	if ( value == NULL && len > 0){
 		cl_log(LOG_ERR, "binary_dup:"
 		       "NULL vlaue with len =%d", 
-		       len);
+		       (int)len);
 		return NULL;
 	}
